PlayerCog: Decompose find_lavalink_track
